:ok_hand: rename <Selector /> to <PreferencePicker />
